requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
coverage==7.8.0
debugpy==1.8.14
decorator==5.2.1
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gotrue==2.12.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.36.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
multidict==6.4.3
mypy==1.16.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
oauthlib==3.2.2
openai==1.82.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.7
pluggy==1.5.0
postgrest==1.0.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
# Removido psycopg2-binary - não precisamos mais do PostgreSQL direto
# psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.7
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
pytest-flask==1.3.0
pytest-mock==3.14.0
pytest-sugar==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyzmq==26.4.0
realtime==2.4.2
redis==6.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
termcolor==3.1.0
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
yarl==1.20.0
